Added site wide card tracking
New DB table "targets" will track the max and average damage done by a
job under card windows. This is a bit of a naive approach because it
doesn't factor in / out a lot of things (pots, dance partner) but for
the moment it works as a proof of concept.