Updated Damage Average Tracking
Updated the average tracking to require the fight to be killed, and for it to distinguish between seperate fights. It now will track all high-end content (designated by FFLogs difficulty of 100+) and will ignore low-end content such as normal trials. Normal vs Savage are all tracked individually as well.

Rework Average Damage in Window Tracking

!!This is a breaking change that will require the targets table of the database to be dropped or reformatted before the calculator will run.

Average tracking now distinguishes between encounters and difficulty. Any non-high end encounter (designated with difficulty 'None' by FFlogs) will be ignored for the total tracking. All high end encounters are now individually tracked. I'll leave it like the for now, but in the future it would be idea to be able to weigh certain windows / ranges by job, especially for the context of Ultimates.