Fixed Report Tracking To Prevent Duplicate Repots
Fixed the report tracking in main.py to prevent duplicate reports from being entered into the database. This should cache
results for the duration of the cardcalc process, and only recaculate if the application has restarted since the original
calculation. I believe this was working before but I broke it when I moved over to PSQL.